+++ src/Actions/MediaAction.php
@@ -3,8 +3,11 @@
 namespace Awcodes\Curator\Actions;
 
 use Awcodes\Curator\Components\Forms\CuratorPicker;
+use Awcodes\Curator\Facades\Curator;
+use Filament\Forms\ComponentContainer;
 use Filament\Forms\Components\Actions\Action;
 use FilamentTiptapEditor\TiptapEditor;
+use Illuminate\Support\Str;
 
 class MediaAction extends Action
 {
@@ -17,6 +20,13 @@ class MediaAction extends Action
     {
         parent::setUp();
 
+        $this->mountUsing(function (TiptapEditor|CuratorPicker $component, ComponentContainer $form) {
+            $src = $component->getLivewire()->mediaProps['src'];
+            $media = $src !== ''
+                ? Curator::getMediaModel()::firstWhere('name', Str::of($src)->afterLast('/')->beforeLast('.'))
+                : null;
+        });
+
         $this->modalWidth('screen');
 
         $this->modalHeading(__('curator::views.panel.heading'));
@@ -24,6 +34,10 @@ class MediaAction extends Action
         $this->modalActions(fn () => []);
 
         $this->modalContent(static function (TiptapEditor|CuratorPicker $component) {
+            $src = $component->getLivewire()->mediaProps['src'];
+            $selected = $src !== ''
+                ? Curator::getMediaModel()::firstWhere('name', Str::of($src)->afterLast('/')->beforeLast('.'))
+                : null;
             return view('curator::components.actions.picker-action', [
                 'statePath' => $component->getStatePath(),
                 'modalId' => $component->getLivewire()->id.'-form-component-action',
@@ -40,6 +54,7 @@ class MediaAction extends Action
                 'imageCropAspectRatio' => app('curator')->getImageCropAspectRatio(),
                 'imageResizeTargetWidth' => app('curator')->getImageResizeTargetWidth(),
                 'imageResizeTargetHeight' => app('curator')->getImageResizeTargetHeight(),
+                'selected' => $selected
             ]);
         });
     }