+++ src/Components/Forms/Uploader.php
@@ -29,7 +29,13 @@ class Uploader extends FileUpload
             $storeMethod = $component->getVisibility() === 'public' ? 'storePubliclyAs' : 'storeAs';
 
             if (Curator::isResizable($extension)) {
-                $image = Image::make($file->getRealPath());
+                if (in_array($file->disk, curator()->getCloudDisks())) {
+                    $content = Storage::disk($file->disk)->get($file->path());
+                } else {
+                    $content = $file->getRealPath();
+                }
+
+                $image = Image::make($content);
                 $image->orientate();
                 $width = $image->getWidth();
                 $height = $image->getHeight();