+++ src/helpers.php
@@ -1,0 +1,10 @@
+<?php
+
+use Awcodes\Curator\Curator;
+
+if (! function_exists('curator')) {
+    function curator(): Curator
+    {
+        return app('curator');
+    }
+}